refactor(pin): simplify window message handling
重构 PinWindow 的消息分发逻辑，将命令处理、默认缩放重置、
右键菜单定位、窗口位置联动和销毁清理拆成独立 helper，降低
HandleMessage 的流程复杂度并保持原有行为不变。

同时清理一批低风险 clang-tidy 告警，包括冗余成员初始化、
MutationScope 的特殊成员约束、日志大小常量的显式类型提升，
以及 pin_window.cpp 中的 ranges 算法替换。

已通过 clang-format、changed clang-tidy 与 debug 预设构建验证；
当前剩余 warning 仅为日志宏定义，后续由单独的 tidy 配置提交处理。

+++ src/core/log.h
@@ -38,7 +38,7 @@ namespace capturezy::core
         LogLevel minimum_level{LogLevel::Info};
         bool debugger_output_enabled{true};
         bool file_output_enabled{true};
-        std::size_t max_file_size_bytes{1024U * 1024U};
+        std::size_t max_file_size_bytes{static_cast<std::size_t>(1024U) * 1024U};
         std::size_t max_retained_files{3};
     };
